Adds attempted GPU odd/even sort.
GPU sort not functional; might not be able to be implemented exclusively in gpu.js, but may be possible if odd/even iteration handled in javascript (only swaps handled on GPU, propagated as textures in for/while loop).

Note: This GPU sort implementation changes return value of getPfIvLevelsSPsAtCpLimit; downstream functions that relied on previous return value may need to be updated to match.